+++ app/Http/Controllers/V1/ImageManipulationController.php
@@ -7,6 +7,7 @@ use App\Http\Requests\ResizeImageRequest;
 use App\Http\Resources\V1\ImageManipulationResource;
 use App\Models\Album;
 use App\Models\ImageManipulation;
+use Illuminate\Http\Request;
 use Illuminate\Http\UploadedFile;
 use Illuminate\Support\Facades\File;
 use Illuminate\Support\Str;
@@ -17,13 +18,17 @@ class ImageManipulationController extends Controller
     /**
      * Display a listing of the resource.
      */
-    public function index()
+    public function index(Request $request)
     {
-        return ImageManipulationResource::collection(ImageManipulation::paginate());
+        return ImageManipulationResource::collection(
+            ImageManipulation::where('user_id', $request->user()->id)->paginate()
+        );
     }
     
-    public function getByAlbum(Album $album)
+    public function getByAlbum(Request $request, Album $album)
     {
+        $this->authorize($request->user()->id, $album->user_id);
+
         $where = [
             'album_id' => $album->id
         ];
@@ -48,11 +53,13 @@ class ImageManipulationController extends Controller
         $data = [
             'type' => ImageManipulation::TYPE_RESIZE,
             'data' => json_encode($all),
-            'user_id' => null
+            'user_id' => $request->user()->id
         ];
 
         if (isset($all['album_id'])) {
-            // TODO authorization
+            $album = Album::find($all['album_id']);
+
+            $this->authorize($request->user()->id, $album->user_id);
 
             $data['album_id'] = $all['album_id'];
         }
@@ -99,16 +106,20 @@ class ImageManipulationController extends Controller
     /**
      * Display the specified resource.
      */
-    public function show(ImageManipulation $image)
+    public function show(Request $request, ImageManipulation $image)
     {
+        $this->authorize($request->user()->id, $image->user_id);
+        
         return new ImageManipulationResource($image);
     }
-
+    
     /**
      * Remove the specified resource from storage.
      */
-    public function destroy(ImageManipulation $image)
+    public function destroy(Request $request, ImageManipulation $image)
     {
+        $this->authorize($request->user()->id, $image->user_id);
+        
         $image->delete();
 
         return response('', 204);
